AMBARI-23404. Customize Services - when user is changed in accounts tab no warning shown to adjust other dependent properties , though shown when navigate to Misc Tab
